requirements.txt
django
djangorestframework
django-redis
django-filter
django-cors-headers
django-filter
django-debug-toolbar
djangorestframework-simplejwt
